Temporarily disable broken tests
