+++ src/app/components/regions/pages/details/details.component.spec.ts
@@ -1,8 +1,8 @@
 import { ComponentFixture, TestBed } from "@angular/core/testing";
-
 import { DetailsComponent } from "./details.component";
 
-describe("DetailsComponent", () => {
+// TODO Implement
+xdescribe("RegionDetailsComponent", () => {
   let component: DetailsComponent;
   let fixture: ComponentFixture<DetailsComponent>;
